# proyecto-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/news": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Muestra un listado de noticias",
                "description": "Devuelve un listado de noticias",
                "operationId": "getArticlesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "añade una noticia",
                "description": "devuelve los datos de una noticia",
                "operationId": "addArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/news/{article}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "obtiene una noticia",
                "description": "devuelve una noticia",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "id de la noticia a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Edita una noticia",
                "description": "devuelve los datos de una noticia",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "admin": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Borra una noticia",
                "description": "Borra una noticia y devuelve un 204",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "admin": []
                    }
                ]
            }
        },
        "/api/news/category/{category}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Muestra un listado de noticias",
                "description": "Devuelve un listado de noticias",
                "operationId": "getArticlesByCategory",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "id de la categoria a mostrar nosticias",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArticleRequest": {
                "title": "Store Article Request",
                "description": "Store Article request body data",
                "required": [
                    "title,description,body,image,categories,user_id"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "titulo de la noticia",
                        "type": "string",
                        "example": "noticia 1"
                    },
                    "description": {
                        "title": "description",
                        "description": "descripcion de la noticia a mostrar brevemente",
                        "type": "string",
                        "example": "noticia sobre..."
                    },
                    "body": {
                        "title": "body",
                        "description": "cuerpo de la noticia",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    },
                    "image": {
                        "title": "image",
                        "description": "nombre del fichero de la imagen de cabecera",
                        "type": "string",
                        "example": "1.jpg"
                    },
                    "categories": {
                        "title": "categories",
                        "description": "array con las categorias a las que pertenece la noticia",
                        "type": "string",
                        "example": [
                            "1",
                            "2",
                            "3"
                        ]
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "id del usuario que ha creado la noticia",
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ArticleRequest"
                }
            },
...
            "ArticleResource": {
                "title": "ArticleResource",
                "description": "Article resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ArticleResource"
                }
            },
...
            "Article": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "titulo de la noticia",
                        "type": "string",
                        "example": "Noticia 1"
                    },
                    "description": {
                        "description": "Descripción de la noticia",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    },
                    "body": {
                        "description": "cuerpo de la noticia",
                        "type": "string",
                        "example": "Lorem ipsum..."
                    },
                    "date": {
                        "description": "fecha de la noticia",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 00:00"
                    },
                    "url": {
                        "description": "url de la noticia",
                        "type": "text",
                        "example": "www.ecample.com"
                    },
                    "image": {
                        "description": "nombre de la imagen de cabecera",
                        "type": "string",
                        "example": "1.jpg"
                    },
                    "user_id": {
                        "description": "id del usuario que ha creado la noticia",
                        "type": "string",
                        "example": "1"
                    },
                    "user_name": {
                        "description": "nombre del usuario que ha creado la noticia",
                        "type": "string",
                        "example": "admin"
                    },
                    "categories": {
                        "description": "categorias a las que pertenece la noticia",
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "example": "23-03-09 18:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de modificación",
                        "type": "string",
                        "example": "23-03-09 18:30:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Article"
                }
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}